backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Autenticación JWT
djangorestframework-simplejwt>=5.3.0

# Base de datos
psycopg2-binary>=2.9.9

# Utilidades
python-dotenv>=1.0.0
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
